+++ include/porto/vec3.h
@@ -11,40 +11,40 @@ namespace porto
 	public:
 		union
 		{
-			float e[3];
+			double e[3];
 			struct
 			{
-				float r;
-				float g;
-				float b;
+				double r;
+				double g;
+				double b;
 			};
 			struct
 			{
-				float x;
-				float y;
-				float z;
+				double x;
+				double y;
+				double z;
 			};
 		};
 
 		// initializer
 		explicit Vec3() : x{0}, y{0}, z{0} {}
-		explicit Vec3(float e0, float e1, float e2) : x{e0}, y{e1}, z{e2} {}
+		explicit Vec3(double e0, double e1, double e2) : x{e0}, y{e1}, z{e2} {}
 		Vec3(const Vec3 &v) : x{v.x}, y{v.y}, z{v.z} {}
 
 		inline const Vec3& operator+() const { return *this; }
 		inline Vec3 operator-() const { return Vec3(-e[0], -e[1], -e[2]); }
-		inline float operator[](int i) const { return e[i]; }
-		inline float& operator[](int i) { return e[i]; }
+		inline double operator[](int i) const { return e[i]; }
+		inline double& operator[](int i) { return e[i]; }
 
 		inline Vec3& operator+=(const Vec3 &v2);
 		inline Vec3& operator-=(const Vec3 &v2);
 		inline Vec3& operator*=(const Vec3 &v2);
 		inline Vec3& operator/=(const Vec3 &v2);
-		inline Vec3& operator*=(const float t);
-		inline Vec3& operator/=(const float t);
+		inline Vec3& operator*=(const double t);
+		inline Vec3& operator/=(const double t);
 
-		inline float length() const { return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]); }
-		inline float squared_length() const { return e[0]*e[0] + e[1]*e[1] + e[2]*e[2]; }
+		inline double length() const { return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]); }
+		inline double squared_length() const { return e[0]*e[0] + e[1]*e[1] + e[2]*e[2]; }
 		inline void normalize();
 	};
 
@@ -59,7 +59,7 @@ namespace porto
 	}
 
 	inline void Vec3::normalize() {
-		float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+		double k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
 		e[0] *= k; e[1] *= k; e[2] *= k;
 	}
 
@@ -75,11 +75,11 @@ namespace porto
 		return Vec3(v1.e[0] * v2.e[0], v1.e[1] * v2.e[1], v1.e[2] * v2.e[2]);
 	}
 
-	inline Vec3 operator*(float t, const Vec3 &v) {
+	inline Vec3 operator*(double t, const Vec3 &v) {
 		return Vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 	}
 
-	inline Vec3 operator*(const Vec3 &v, float t) {
+	inline Vec3 operator*(const Vec3 &v, double t) {
 		return Vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 	}
 
@@ -87,11 +87,11 @@ namespace porto
 		return Vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 	}
 
-	inline Vec3 operator/(Vec3 v, float t) {
+	inline Vec3 operator/(Vec3 v, double t) {
 		return Vec3(v.e[0]/t, v.e[1]/t, v.e[2]/t);
 	}
 
-	inline float dot(const Vec3 &v1, const Vec3 &v2) {
+	inline double dot(const Vec3 &v1, const Vec3 &v2) {
 		return v1.e[0]*v2.e[0]
 			 + v1.e[1]*v2.e[1]
 			 + v1.e[2]*v2.e[2];
@@ -124,7 +124,7 @@ namespace porto
 		return *this;
 	}
 
-	inline Vec3& Vec3::operator*=(const float t) {
+	inline Vec3& Vec3::operator*=(const double t) {
 		e[0] *= t;
 		e[1] *= t;
 		e[2] *= t;
@@ -138,8 +138,8 @@ namespace porto
 		return *this;
 	}
 
-	inline Vec3& Vec3::operator/=(const float t) {
-		float k = 1.0/t;
+	inline Vec3& Vec3::operator/=(const double t) {
+		double k = 1.0/t;
 		e[0] *= k;
 		e[1] *= k;
 		e[2] *= k;
